Missing escape chars in healthcheck variables
Added missing $ to escape the variables so it can be expanded inside the container. Otherwise the healtcheck outputs and error and docker considers the variables as blank.

$ sudo docker-compose up
WARNING: The POSTGRES_USER variable is not set. Defaulting to a blank string.
WARNING: The POSTGRES_DB variable is not set. Defaulting to a blank string.

Logs:
postgres_1    | 2024-11-30 01:36:40.966 UTC [917] FATAL:  role "-d" does not exist