Update init-data.sh
The current version of the script creates the PostgreSQL user in lowercase only. This happens because the variable POSTGRES_NON_ROOT_USER is not quoted properly in the SQL command. As a result, any uppercase letters in the username are ignored, and the user is created in lowercase.
This leads to connection issues when n8n tries to connect to PostgreSQL using the original username with mixed case, since PostgreSQL treats unquoted identifiers as lowercase by default.

before:
${POSTGRES_NON_ROOT_USER}

now:
"${POSTGRES_NON_ROOT_USER}"